requirements.txt
absl-py==1.3.0
aio-pika==8.2.3
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.4
aiormq==6.4.2
aiosignal==1.3.1
appnope==0.1.3
APScheduler==3.9.1.post1
astroid==2.15.4
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.12.2
bidict==0.22.1
boto3==1.26.133
botocore==1.29.133
bs4==0.0.1
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
comm==0.1.3
confluent-kafka==1.9.2
cryptography==40.0.2
cycler==0.11.0
dask==2022.10.2
debugpy==1.6.7
decorator==5.1.1
dill==0.3.6
dnspython==1.16.0
docopt==0.6.2
executing==1.2.0
fbmessenger==6.0.0
fire==0.5.0
flatbuffers==23.5.9
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.3
gast==0.4.0
google-auth==2.18.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.2
h5py==3.8.0
httptools==0.5.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.22.0
ipython==8.13.1
isort==5.12.0
jedi==0.18.2
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.1
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
keras==2.11.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
libclang==16.0.0
locket==1.0.0
lxml==4.9.2
magic-filter==1.0.9
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mattermostwrapper==2.2
mccabe==0.7.0
msgpack==1.0.5
multidict==5.2.0
nest-asyncio==1.5.6
networkx==2.6.3
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==20.9
pamqp==3.2.1
pandas==2.0.1
parso==0.8.3
partd==1.4.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
pluggy==1.0.0
portalocker==2.7.0
prompt-toolkit==3.0.28
protobuf==3.19.6
psutil==5.9.5
psycopg2==2.9.6
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.2
pydot==1.4.2
Pygments==2.15.1
PyJWT==2.7.0
pykwalify==1.8.0
pylint==2.17.3
pymongo==3.10.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-engineio==4.4.1
python-socketio==5.8.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.2
questionary==1.10.0
randomname==0.1.5
rasa==3.5.8
rasa-sdk==3.5.1
redis==4.5.5
regex==2022.10.31
requests==2.29.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rocketchat-API==1.28.1
rsa==4.9
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.8.1
sentry-sdk==1.14.0
six==1.16.0
sklearn-crfsuite==0.3.6
slack-sdk==3.21.3
soupsieve==2.4.1
SQLAlchemy==1.4.48
stack-data==0.6.2
tabulate==0.9.0
tarsafe==0.0.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-addons==0.19.0
tensorflow-estimator==2.11.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-text==2.11.0
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tomli==2.0.1
tomlkit==0.11.8
toolz==0.12.0
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
twilio==7.14.2
typeguard==4.0.0
typing-utils==0.1.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
ujson==5.7.0
Unidecode==1.3.6
urllib3==1.26.15
uvloop==0.17.0
wcwidth==0.2.6
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==2.3.4
wrapt==1.15.0
yarl==1.9.2
zipp==3.15.0
